+++ CHANGELOG.md
@@ -5,8 +5,28 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2024-03-16
+## [2.0.0] - 2024-03-18
+### Added
+- N/A
 
+### Changed
+- Renamed `Aurora::make` -> `Aurora::create`
+
+### Deprecated
+- N/A
+
+### Removed
+- N/A
+
+### Fixed
+- N/A
+
+### Security
+- Added additional unit tests
+
+---
+
+## [1.0.0] - 2024-03-16
 ### Added
 - Initial release of Aurora DB.
 - Implemented the `Aurora` model for interacting with database tables.